requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pillow>=10.0.0
azure-functions>=1.14.0
azure-storage-blob>=12.17.0
azure-cognitiveservices-vision-computervision>=0.9.0
python-docx>=0.8.11
PyPDF2>=3.0.0
plotly>=5.15.0
streamlit-option-menu>=0.3.6
python-dotenv>=1.0.1
